migrate _read_events to utils and change min_duration to 0
